pyproject.toml
[project]
name = "guidely-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "fastapi>=0.118.0",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.10",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.8",
    "langgraph-cli[inmem]>=0.4.2",
    "langsmith>=0.4.31",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "streamlit>=1.50.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.37.0",
]
